pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "distillgov"
version = "0.1.0"
description = "Congress, distilled. Making government accessible to everyone."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Data ingestion
    "httpx>=0.27.0",
    "capitolgains>=0.1.0",
    "playwright>=1.41.0",

    # Database
    "duckdb>=1.0.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
    "typer>=0.13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "pyright>=1.1.0",
]
pipeline = [
    "apache-airflow>=2.10.0",
    "dbt-core>=1.8.0",
    "dbt-duckdb>=1.8.0",
]

[project.scripts]
distill = "ingestion.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
